First stab at properly supporitng torch.dtype
test/test_jit*.py seems to pass

Including subsystem updates to make the respective tests pass:
- dtype updates in cuda fuser
- making the tensorexpr fuser work, fix more tests
- ONNX vs. JIT dtype support

+++ torch/csrc/jit/frontend/function_schema_parser.cpp
@@ -191,11 +191,11 @@ struct SchemaParser {
         auto tok = L.next();
         auto text = tok.text();
         if ("float" == text) {
-          return static_cast<int64_t>(at::kFloat);
+          return at::kFloat;
         } else if ("complex" == text) {
-          return static_cast<int64_t>(at::kComplexFloat);
+          return at::kComplexFloat;
         } else if ("long" == text) {
-          return static_cast<int64_t>(at::kLong);
+          return at::kLong;
         } else if ("strided" == text) {
           return static_cast<int64_t>(at::kStrided);
         } else if ("Mean" == text) {
